backend/requirements.txt
# 수정일: 2026-02-24
# 수정내용: Django 백엔드 핵심 라이브러리 정의 및 Socket.io 통신 관련 중복 라이브러리 제거

django
gunicorn>=21.2.0
django-environ
psycopg2-binary
djangorestframework
django-cors-headers
drf-spectacular
dj-database-url
datasets
deep-translator
openai
python-dotenv
Pillow
google-genai
google-generativeai
groq
boto3
django-storages

# Job Planner Agent
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
# sentence-transformers>=2.2.0
# torch>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=2.0.0
playwright>=1.40.0
google-api-python-client==2.116.0


# LangGraph Agent
langgraph>=0.2.0
langchain-core>=0.3.0

# Real-time Collaboration (Socket.io)
python-socketio>=5.11.0
python-engineio>=4.9.0
uvicorn>=0.27.0

